VDB.Config

Description

Configurations.

Documentation

type Config b = Feature -> b

A configuration is a function that indicates whether each feature is enabled (true) or disabled (false). The boolean return value is parameterized to admit alternative logics or symbolic values.

enableAll :: Boolean b => Config b

A configuration that enables all features.

disableAll :: Boolean b => Config b

A configuration that disables all features.

enable :: Boolean b => Feature -> Config b -> Config b

Override a configuration to enable an option.

disable :: Boolean b => Feature -> Config b -> Config b

Override a configuration to disable an option.

enableMany :: Boolean b => [Feature] -> Config b -> Config b

Override a configuration to enable the indicated features.

disableMany :: Boolean b => [Feature] -> Config b -> Config b

Override a configuration to disable all except the indicated features.