VDB.FeatureExpr

Description

Feature expressions.

Documentation

data FeatureExpr

Boolean expressions over features.

Constructors

Lit Bool

Ref Feature

Not FeatureExpr

And FeatureExpr FeatureExpr

Or FeatureExpr FeatureExpr

features :: FeatureExpr -> Set Feature

The set of features referenced in a feature expression.

evalFeatureExpr :: Boolean b => Config b -> FeatureExpr -> b

Evaluate a feature expression against a configuration.

selectFeatureExpr :: Feature -> Bool -> FeatureExpr -> FeatureExpr

Select a feature within a feature expressions, potentially simplifying it.

prettyFeatureExpr :: FeatureExpr -> String

Pretty print a feature expression.

symbolicFeatureExpr :: FeatureExpr -> Predicate

Generate a symbolic predicate for a feature expression.

shrinkFeatureExpr :: FeatureExpr -> FeatureExpr

Reduce the size of a feature expression by applying some basic simplification rules.