VDB.SAT

Description

A simple SAT solving interface.

TODO: need interface for incremental SAT (see: Data.SBV.Control)

Documentation

class Boolean b => SAT b where

A type class for types that can be converted to symbolic predicates and checked by a SAT solver.

Minimal complete definition

toPredicate

Methods

toPredicate :: b -> Predicate

satisfiable :: SAT b => b -> Bool

Is the predicate satisfiable?

unsatisfiable :: SAT b => b -> Bool

Is the predicate unsatisfiable?

tautology :: SAT b => b -> Bool

Is the predicate a tautology?

equivalent :: SAT b => b -> b -> Bool

Are these predicates equivalent?

Orphan instances

Boolean b => Boolean (Symbolic b)

Methods

true :: Symbolic b

false :: Symbolic b

bnot :: Symbolic b -> Symbolic b

(&&&) :: Symbolic b -> Symbolic b -> Symbolic b

(|||) :: Symbolic b -> Symbolic b -> Symbolic b

(~&) :: Symbolic b -> Symbolic b -> Symbolic b

(~|) :: Symbolic b -> Symbolic b -> Symbolic b

(<+>) :: Symbolic b -> Symbolic b -> Symbolic b

(==>) :: Symbolic b -> Symbolic b -> Symbolic b

(<=>) :: Symbolic b -> Symbolic b -> Symbolic b

fromBool :: Bool -> Symbolic b