VDB.Schema

Description

Variational database schemas.

Documentation

type RowType = [Opt (Attribute, Type)]

Type of a relation in the database.

type Schema = Opt (Map Relation (Opt RowType))

A schema is a mapping from relations to row types. Both the map itself and each row type are optionally included. The top-level Opt corresponds to the feature model, which defines the set of valid configurations.

featureModel :: Schema -> FeatureExpr

The feature model associated with a schema.

lookupRowType :: Relation -> Schema -> Maybe (Opt RowType)

Get the type of a relation in the database.