VDB.Variational

Description

A generic interface for variational things.

Optional values

type Opt a = (FeatureExpr, a)

An optionally included value. Whether it is included is determined by a presence condition.

selectOpt :: Feature -> Bool -> Opt a -> Opt a

Apply a selection to an optional value.

configureOpt :: Config Bool -> Opt a -> Maybe a

Apply a configuration to an optional value.

configureOptList :: Config Bool -> [Opt a] -> [a]

Configure a list of optional values.

Variational type class

class Variational a where

A type class for variational things.

Minimal complete definition

choice, choiceMap

Methods

choice :: FeatureExpr -> a -> a -> a

Create a choice.

choiceMap :: (FeatureExpr -> a -> a -> a) -> a -> a

Map a function over all choices.

select :: Feature -> Bool -> a -> a

Partially configure a variational thing.

configure :: Config Bool -> a -> a

Fully configure a variational thing.

selectHelper :: Variational a => Feature -> Bool -> FeatureExpr -> a -> a -> a

A helper function to implement select for data types with simple choices.

configureHelper :: Config Bool -> FeatureExpr -> a -> a -> a

A helper function to implement configure for data types with simple choices.